+++ src/features/swap/buildTx.ts
@@ -4,15 +4,16 @@ import {
   type TransactionResult,
 } from "@mysten/sui/transactions";
 import BigNumber from "bignumber.js";
-import { getSplitCoinForTx } from "./libs/getSplitCoinForTx";
-import { groupSwapRoutes } from "./libs/groupSwapRoutes";
-import { swapWithRoute } from "./libs/swapWithRoute";
-import { denormalizeTokenType } from "./utils/token";
-import { SuiUtils } from "./utils/sui";
-import type { BuildTxParams } from "./types/tx";
-import { _7K_CONFIG, _7K_PACKAGE_ID, _7K_VAULT } from "./constants/_7k";
+import { MATH_PACKAGE_ID } from "../../constants/math";
+import { getSplitCoinForTx } from "../../libs/getSplitCoinForTx";
+import { groupSwapRoutes } from "../../libs/groupSwapRoutes";
+import { swapWithRoute } from "../../libs/swapWithRoute";
+import { denormalizeTokenType } from "../../utils/token";
+import { SuiUtils } from "../../utils/sui";
+import { BuildTxParams } from "../../types/tx";
+import { _7K_CONFIG, _7K_PACKAGE_ID, _7K_VAULT } from "../../constants/_7k";
 import { isValidSuiAddress } from "@mysten/sui/utils";
-import { MATH_PACKAGE_ID } from "./constants/math";
+import { getConfig } from "./config";
 
 export const buildTx = async ({
   quoteResponse,
@@ -107,6 +108,7 @@ export const buildTx = async ({
   }
 
   const coinObjects: TransactionObjectArgument[] = [];
+  const config = await getConfig();
   await Promise.all(
     routes.map(async (route, index) => {
       const inputCoinObject = coinData[index];
@@ -115,6 +117,7 @@ export const buildTx = async ({
         inputCoinObject,
         currentAccount: accountAddress,
         tx,
+        config,
       });
       if (coinRes) {
         coinObjects.push(coinRes);